# Optimize `_key` to prevent string allocation when formatting `Symbol`s