@moberegger 2025-06-03
@rafaelfranca We've observed massive savings in memory allocation from this change alone. We're talking upwards of hundreds of thousands of strings allocations per minute!

Figured I'd let you know in case you needed some convincing... and also to show off a bit 😆.